Järjestelty HakuFilters-komponentin koodia paremmin
