+++ src/app/lib/common.ts
@@ -36,3 +36,35 @@ export class FetchError extends Error {
     this.response = response;
   }
 }
+
+function isObject(value: unknown): value is Record<string, unknown> {
+  return typeof value === 'object' && value !== null;
+}
+
+function isTranslatedName(value: unknown): value is TranslatedName {
+  return (
+    isObject(value) &&
+    (typeof value?.fi === 'string' ||
+      typeof value?.sv === 'string' ||
+      typeof value?.en === 'string')
+  );
+}
+
+export const byProp = <
+  T extends Record<string, string | number | TranslatedName>,
+>(
+  key: string,
+  direction: 'asc' | 'desc' = 'asc',
+  lng: Language,
+) => {
+  const asc = direction === 'asc';
+  return (a: T, b: T) => {
+    const aKey = a[key];
+    const aProp = isTranslatedName(aKey) ? getTranslation(aKey, lng) : aKey;
+
+    const bKey = b[key];
+    const bProp = isTranslatedName(bKey) ? getTranslation(bKey, lng) : bKey;
+
+    return aProp > bProp ? (asc ? 1 : -1) : bProp > aProp ? (asc ? -1 : 1) : 0;
+  };
+};
